Reorganize project, writing, ...
The project and API have been reorganized, but it will not be the last time it
will be done. The non-blocking aspect works now, apart from a few corner cases.
The write a file to a server function needs to be implemented now.

'main' has been changed so it is now usable and accepts arguments instead of
hard coded strings.

Debugging has been improved with the states description now printed out.

socket and file handles are now type-defined if they need changing.