Initialization state, server outline, bind, make
CppCheck has been added as a Makefile target, and the makefile changed a
little.

More work is done in the initialization state of a TFTP transaction, this
should aid in code reuse between the server and the client modes.

An outline of the server has been added which is non-functional at the moment.

'bind' has been added as an option, and the nopen extended to handle opening
ports for servers.