Added non-working server
The server has been added, but it does not work at the moment which is to be
expected - the code has just been added and there is a lot to go wrong. The
callback API has been modified to accommodate it, but the state machine has
not, which will be the next step.

Makefile targets for the server have also been added.