Server talking, partially working, reorg
The structure of the project has been reorganized and improved, and a few bugs
have been fixed - there is still a lot more to do.

Here are some of the changes in this commit:

- 'pack.md', a plan for a serialization compiler has been added
- The RFC for TFTP has been moved to a sub directory
- An outline for the Windows OS dependent code has been added
- Functions are now stored globally in a table in the OS dependent code file
instead of per instance.
- Some of the server code has been split off into separate functions
for readability.
- Some bugs relating to initialization and cleanup have been fixed,
many other bugs remain however, so it is not too clear what has been
fixed. The client seems to work okay, with some failure modes relating to block
numbers and time outs, the server basically does not work - but now it responds
which is a step up.