Some Windows stuff, basic server working
The basic server functionality seems to be working, this is due to a
combination of bug fixes and subtle server dependent changes to how things
work.

The MingW tool chain can build 'win.c' and 'tftp.c' under Windows and some of
the Windows OS code has been filled out, it is not completed yet.