Added log levels, packet processor, ...
There are quite a few changes in this commit, most of them are systematic
changes introduced by the new log-level mechanism which provides a fairly
flexible logging facility to the project.

Three packet processors have been added, the build system can be changed so
that any of the three can be used without any changes to the source code as
they all use the same interface. The processor can modify or analyze the
packets sent and received by the TFTP server/client.

The packet processor modules are:
* 'print.c', which performs a hexdump of the packet
* 'null.c', which does nothing with the packet, this is default
* 'error.c', which introduces random errors for testing

Other changes include some whitespace, more documentation and moving things
around to support the packet processor functionality.