Windows, error injector
The main changes in this commit are improvements to the Windows port, which
although it is not fully functional it does now send/receive packets (just not
multiple ones) in both client and server modes.

The error injector has been much improved, it now has nearly fully configurable
behavior with a better random number generator. There are a few improvements to
be made, but this can be used to generate data for testing purposes now.